requirements.txt
asgiref==3.7.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-health-check==3.17.0
djangorestframework==3.14.0
django_filter==23.2
drf-yasg==1.21.6
firebase_admin==6.2.0
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0
requests~=2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
